Blink: Make wallet only long poll blinks
- Sometimes the wallet would query TX's before they are blink approved
and get cached into the wallet as a normal TX preventing external
services from using 0-confirmation on transactions.

We make the wallet only conduct long-polling for blinks in order to
notice the transaction after blink. The downside here is that normal
mempool transacations won't appear until they are placed into a block.

With Pulse this is mitigated with reliable within sub-seconds of
a target 2 min block time.