wallet: Re-sort pooled blink outputs when confirmed
Due to prior behaviour of immediately accepting blinks, they get
commited to the wallet's transfers container early. A second view wallet,
that restores the first wallet seed will _not_ see the blinks if it is
generated after the fact as this information is not stored in the
blockchain.

The second view wallet will restore transactions from the blockchain in
the order they were committed to the blockchain (oblivious to any blink
transactions as this is not stored by the blockchain, only those who
see it in the mempool at the time it is relayed would optimistically
save it to their transfers container).

This leads to importing key images breaking as the two wallets expect
_slightly_ different key images at different indexes in their respective
transfer container.

The fix here is to sort all transfers again once we confirm the blink,
placing the transfer into the correct slot in the container. This sort
only occurs within the latest block (as blinks are only relevant at
the tip of the chain) so resorting occurs typically with 1 blocks worth
of transactions for the wallet and so will always be quick and not
expensive.

You can reproduce this bug by checking out the tag v8.1.1

1. Wallet A sends to Wallet B transfer_1 without blink `transfer unimportant WalletB <amount>`
2. Wallet A sends to Wallet B transfer_2 _before_ transfer_1 gets mined into a block `transfer blink WalletB <amount>`
3. Wallet B must receive the Blink and be refreshed _before_ the next block gets mined to commit it to Wallet B's cache.
4. Wait for the next block to get mined and refresh to accept the block

- Wallet B now has instantly confirmed transfer_2 the Blink into the transfer
  container at index 0.

- Wallet B confirms the slow transfer transfer_1 and saves it into transfer
  container at index 1.

Since the network received transfer_1 first in the mempool, it will get
included in the block first with the Blink second. Now Wallet B's
transfers are out of sync with what is visible on the network.

5. export_key_images and import it into a view only wallet based on Wallet B.
6. import_key_images will fail due to the blockchain/wallet cache mismatch.