Error when trying to sign using a watch-only wallet
The cli wallet was correctly refusing to sign, but the RPC wallet (and
wallet2 internally) allowed it and would sign with the all-0s address,
which of course is invalid and fails to verify.  This now throws if
attempting to do so to get an error returned to the rpc endpoint.

Other minor changes here:
- make wallet2::verify static (verification doesn't depend on the wallet
  instance at all).
- Add documentation for wallet2 sign/verify
- Slight DRY of wallet2::sign