Add helpful message to error string
A few times I've connected my test Ledger but forgotten to open the app,
and then had connection failures without any indicating of what went
wrong (just "No device found").  This adds "(Is the device running with
the wallet app opened?)" to that error message as that is likely often
the culprit.