Retrofit versioning to state change serialization
This reinterprets the leading "state" as a version field, if >= 4, and
otherwise keeps it as the state value if < 4.

It is done in such a way as to remain the same round-trip (i.e. if we
deserialize it and then reserialize) so as to not break existing
signature verification.

This lets us properly serialize/deserialize both old, reasonless state
change txes *and* new state change txes with a reason field.  Without
this syncing failed because we'd hit a state change tx and couldn't
parse it properly.  (But we also can't just "upgrade" to the new version
because that would change the serialized value and break signatures).