Fix wrong change value in outputs sent to self
We figure out, when parsing a blockchain-confirmed transaction, what the
change is but never properly set that in the new
confirmed_transfer_details entry that we create from the
unconfirmed_transfer_details, and the latter only has the *actual*
change but doesn't know about outputs we sent to ourself.

This commit fixes it by updating the change value when such a case
occurs, which fixes a wrong output & running total bug in csv exports.