Fix string construction when no pow_hash
This was throwing when there was no pow_hash (can't construct a string
from a nullptr).

It doesn't make much sense to stick an empty string there at all,
though, so this amends it to only set the `pow_hash` key when there is
something to set (rather than setting it to an empty string).